requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
apihelper==0.1.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==2.1.1
Deprecated==1.2.13
distro==1.8.0
Django==4.2.7
django-crontab==0.7.1
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.4.0
idna==3.6
importlib-metadata==6.0.0
loguru==0.7.2
magic-filter==1.0.12
Markdown==3.4.1
multidict==6.0.4
netaddr==0.8.0
openpyxl==3.1.2
Pillow==9.5.0
psycopg2==2.9.7
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
tinydb==4.8.0
typing_extensions==4.8.0
urllib3==1.26.14
var-dump==1.2
wrapt==1.14.1
yarl==1.9.3
yookassa==2.3.4
zipp==3.12.0
